Fix GitHub Pages DataCloneError: gate the P2P worker on crossOriginIsolated
On GitHub Pages there are no COOP/COEP response headers, so the page is NOT
cross-origin isolated on first load — the coi-serviceworker installs and
reloads the page to make it isolated. But p2p.js constructed the wasm prover
Worker and shared a SharedArrayBuffer with it at module-load time, on that
first not-yet-isolated load, throwing:

  Failed to execute 'postMessage' on 'Worker': SharedArrayBuffer transfer
  requires self.crossOriginIsolated

(Locally this never reproduced: vite dev/preview send COOP/COEP, so the very
first load is already isolated.)

Split the entry: p2p.js is now a tiny loader that dynamically imports the
real bootstrap (p2p-boot.js, the former p2p.js verbatim) only when
self.crossOriginIsolated is true; on a non-isolated first load it shows a
"reloading" notice and touches no wasm/worker, letting the service worker
reload into the isolated context.

Verified against a faithful Pages replica (static server, no COOP/COEP):
no DataCloneError, the SW reloads, and the app self-heals to isolated with
the worker up (transport=trystero). The already-isolated path (vite preview)
still connects peers over trystero.
